requirements.txt
streamlit>=1.40.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
chromadb>=0.5.0
python-dotenv>=1.0.0
